split configuration of rspec / minitest into separate files
