+++ lib/match_json/minitest.rb
@@ -1,0 +1,7 @@
+require 'match_json'
+
+class Minitest::Test
+  include MatchJson::Matchers
+
+  alias_method :match_json, :include_json
+end
