+++ test/test_helper.rb
@@ -1,5 +1,5 @@
 gem 'minitest'
 require 'minitest/autorun'
 require 'minitest/matchers'
-require 'match_json'
+require 'match_json/minitest'
 require 'minitest/byebug' if ENV['DEBUG']
